refactor: simplify safe apps to single custom apps view
